Updated databasecontext
